Change doc return type

+++ metrics/ast.py
@@ -349,7 +349,7 @@ def varcomp(tlist: list[tuple[Any, javalang.tree.ClassDeclaration]]) -> float:
 
 def pcn(tlist: list[tuple[Any, javalang.tree.ClassDeclaration]]) -> int:
     """Return number of words in the name of a class.
-    r:type: float
+    r:type: int
     """
     classname = tlist[0][1].name
     # By naming convention Java classes names use PascalCase.
